+++ CHANGELOG.rst
@@ -41,6 +41,11 @@ Fixed
 
   `Launchpad-bug #1881456 <https://launchpad.net/bugs/1881456>`_
 
+Removed
+~~~~~~~
+
+* Remove deprecated api version 1 for designate
+
 [2.0.0] - 2020-05-08
 --------------------